requirements.txt
requests
py-cord
dadjokes
box
aiohttp
Pillow